Remove duplicate logic, create useURLControl hook
