+++ src/store/markets/useSearchMarketsURLControl.ts
@@ -1,7 +1,6 @@
-import { useEffect } from 'react'
-
-import { useLocation, useNavigate } from 'react-router-dom'
 import { create } from 'zustand'
+
+import { useURLControl } from './helpers'
 
 type SelectedMarketsStore = {
   selectedMarkets: string[]
@@ -35,23 +34,9 @@ const useSelectedMarkets = create<SelectedMarketsStore>((set) => {
 })
 
 export const useSearchMarketsURLControl = () => {
-  const navigate = useNavigate()
-  const location = useLocation()
-
   const { selectedMarkets, setSelectedMarkets } = useSelectedMarkets()
 
-  useEffect(() => {
-    const queryParams = new URLSearchParams(location.search)
-    queryParams.delete('selectedMarkets')
-
-    if (selectedMarkets.length > 0) {
-      queryParams.append('selectedMarkets', selectedMarkets.join(','))
-    }
-
-    navigate({ search: queryParams.toString() })
-
-    localStorage.setItem(LOCAL_STORAGE_KEY, JSON.stringify({ selectedMarkets }))
-  }, [selectedMarkets, navigate, location.search])
+  useURLControl({ key: 'selectedMarkets', data: selectedMarkets, storageKey: LOCAL_STORAGE_KEY })
 
   return { selectedMarkets, setSelectedMarkets }
 }
